// src/app/models/convenios.model.ts
export class RegistroConvenios {
    titulo: any;
    organo_responsable: any;
    desde_description_organo: any;
    hasta_description_organo: any;
    num_convenio_1: any;
    num_convenio_2: any;
    num_convenio_3: any;
    num_convenio_4: any;
    desde_importe_gastos: any;
    hasta_importe_gastos: any;
    desde_importe_ingresos: any;
    hasta_importe_ingresos: any;
    organos_firmantes: any;
    entidades_firmantes: any;
    tipo_entidades: any;
    tipo_convenio: any;
    materia: any;
}